Raw lines must now start with .raw
This was primarily done because raw lines were commony used to make inline
data tables.  However, without some denotation to identify the line was
raw, it was unclear when encountering a label as the first thing on a line
whether we're defining a new label, accidentally redifining an existing
label, or whether we intended to store the address represented ny the label
at the address of the line.